Deprecate previous teachers API endpoint and update endpoint to optimized version
