+++ server/routers/teachersRouter.js
@@ -1,8 +1,11 @@
-import express from 'express';
+import express from 'express'
 import * as teachersService from '../services/teachersService.js';
 const teachersRouter = express.Router();
 
-teachersRouter.route('/:id')
-  .get(teachersService.retrieve);
+teachersRouter.route('/home/:id')
+  .get(teachersService.retrieveHome);
+
+teachersRouter.route('/course/:id')
+  .get(teachersService.retrieveCourse);
 
 export default teachersRouter;
